+++ files/private-chef-cookbooks/private-chef/attributes/default.rb
@@ -25,6 +25,12 @@ default['private_chef']['user']['shell'] = "/bin/sh"
 # The home directory for the chef services user
 default['private_chef']['user']['home'] = "/opt/opscode/embedded"
 
+###
+# bootstrap
+###
+default['private_chef']['bootstrap']['admin_email'] = "kryptonite@opscode.com"
+default['private_chef']['bootstrap']['admin_password'] = "kryptonite"
+
 ####
 # CouchDB
 ####